Split out node-based List implg
